api/requirements.txt
anyio==3.6.2
banana-dev==4.0.1
beautifulsoup4==4.11.2
black==23.1.0
bs4==0.0.1
certifi==2022.12.7
langchain
charset-normalizer==3.0.1
click==8.1.3
dnspython==2.3.0
fastapi==0.89.1
h11==0.14.0
idna==3.4
loguru==0.6.0
lxml==4.9.2
mypy-extensions==1.0.0
nanoid==2.0.0
numpy==1.24.1
packaging==23.0
pandas==1.5.3
pathspec==0.11.0
pinecone-client==2.1.0
platformdirs==2.6.2
pydantic==1.10.4
pymongo==4.3.3
python-dateutil==2.8.2
pytz==2022.7.1
PyYAML==6.0
requests==2.28.2
six==1.16.0
sniffio==1.3.0
soupsieve==2.3.2.post1
starlette==0.22.0
tomli==2.0.1
tqdm==4.64.1
typing_extensions==4.4.0
urllib3==1.26.14
uvicorn==0.20.0
